# Conflicts:
#	blockchainapp.xcodeproj/project.pbxproj
#	blockchainapp/ChannelViewController.swift
#	blockchainapp/ChannelsEmitter.swift
#	blockchainapp/FeedBuilders.swift
#	blockchainapp/FeedViewModel.swift
#	blockchainapp/Profile/Likes/LikesEmitter.swift
#	blockchainapp/Profile/Profile/ProfileEmitter.swift
#	blockchainapp/Profile/ProfileBuilder.swift
#	blockchainapp/Profile/ProfileViewController.swift
#	blockchainapp/SearchPresenter.swift
#	blockchainapp/SearchViewController.swift
#	blockchainapp/TrackViewModel.swift
#	blockchainapp/models/MainModels.swift
#	blockchainapp/scenes/Channels/ChannelsViewController.swift
#	blockchainapp/scenes/Feed/FeedViewController.swift

+++ blockchainapp/Station.swift
@@ -54,7 +54,7 @@ class Station: Object {
 	@objc dynamic var image: String = ""
 	@objc dynamic var sourceURL: String = ""
 	@objc dynamic var subscriptionCount: Int = 0
-	@objc dynamic var trackCount: Int = 0
+	@objc dynamic var trackCount: Int64 = 0
 	@objc dynamic var lang: String     		= ""
 	var tags: List<Tag> = List<Tag>()
